libnixf: extract parseInterpolation func, rename parseInterpolableParts
parseInterpolableParts -> parseStringParts

`StringParts` are different from path fragments because paths have no *surrounding quotes*.
So, the first path fragment should be parsed normally, and the rest in "PS_Path" context.
Previously `parseInterpolableParts` was used for only unified parsing contexts (and this only works for strings).
So it was renamed to `parseStringParts`.